Cleaned up and documented SpanCompatibility
