+++ CueSheetNet/CueSheetNet.csproj
@@ -27,10 +27,6 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Folder Include="Compatibility\" />
-	</ItemGroup>
-
-	<ItemGroup>
 	  <Using Include="System.Text" />
 	</ItemGroup>
